Create locale_check.py
Initial commit. As long as I cannot really contribute in Tuxemon development, I prefer to do something related to translations. 
This script is a bit useful to easily check wether the translation is up-to-date. By now it can be done manually, but in the future maybe this script could be more useful. As of now, it just check online translations. It could be improved by implementing file reading, in case of ongoing offline translation. Every function is explained, one is not working, maybe because I'm stupid.
Simply reject this file if you don't find it useful at all! At least I did some exercise in Python ;)