Revamp professor's encounter
This will now be the first of two or three. Still need to do stuff elsewhere (events and such), but it's coming together.

Also, I fully expect a rewrite of this portion as it seems to not flow as well as it should (at least to me). Might be better with events in the map.