Update main storyline (#824)
* Add Allie sprites

* Update allie
whoops

* I just talked about how I don't do mini commits
-___-

Sorry guys

* Add Bob and xerogrund1 to Taba Town map
The first NPCs!

* Time to rewrite everything
Finally :D

* Revamp professor's encounter

This will now be the first of two or three. Still need to do stuff elsewhere (events and such), but it's coming together.

Also, I fully expect a rewrite of this portion as it seems to not flow as well as it should (at least to me). Might be better with events in the map.

* Pretty up the encounter along route 20

* Finish the first encounter with the professor
It's done! Granted, it's not a lot, but it flows well now.

* Create second announcement

Holy Text Block Batman! Yes, it's a lot, but I feel it's too important to split.

* Tying up a loose end

It was bugging me, so I changed it

* Better clarify what The Trial is

* Rearrange so it's easier to read

* Make it so that you can't leave Taba Town yet

* Get everything set up for the announcement

* There's something really wrong with this...

* Spelling fix

Co-authored-by: ZhongQian TiaoGong <sassafrass@fasizzle.com>
Co-authored-by: Leif Theden <leif.theden@gmail.com>