Next step in the story (#856)
* Update story and fix issues with PRs

* Get some variables working and introduce rocky
He's a rock!

* This fixes the pr

* Fix it so that Tiled doesn't scream

* Finish the story for getting a tuxemon

Now to edit the events in tiled to make it all work

* Fix typos

* Fix small issues

* Set up graphics properly
This took a lot longer than it should have

* Split these so I can have events between them

* Create professor

* clamp the sides and continue story

* Continue with story
Slowly...

* Continue story
Proofing and give a reason why the professor is in their house. He's not a creep...

* Go the registration papers!

* Continue story
Almost there...

* Continue story
Almost there....
Battles should be functional here, but they aren't

* Continue story
Tried to balance out the professor's tuxemon... nope, didn't work

* You can now leave the lab! (sorta)
fixed the story some so the postboy leaves

* Continue Story
A Milestone! You can now leave the town (with a tuxemon)

* Update professor's lab

* Switch to npc_move so NPCs don't fly everywhere

* Fix by @xirsoi
NPCs can have tuxemon now!

Co-authored-by: ZhongQian TiaoGong <sassafrass@fasizzle.com>
Co-authored-by: zetty1 <zett@hello.com>