Add actions to lock and unlock controls.
This is done by creating a state that discards events going to lower
states in the stack. This allows the user to still interact with dialogs
created after the controls have been locked.

I have tested this adding the actions manually to the initial
taba_town.tmx cutscene.

Closes #1110. This is an alternative to #840 that is more general, as it
will even work with states that are not the WorldState (for example,
minigames).