Adding new features in tox.ini
The scripts/fmt.sh script has been replaced with a new tox command: tox -e fmt.
It behaves exactly the same way, except that it's executed in a venv.

Another new command, tox -e style, also checks for formatting mistakes.
It returns a code error, if an error has been found in one of the files.
It does not edit any files. It would be suitable for a CI.

The file Makefile has been modified so that "make format" still works.