Number_or_variable: fix crash when passing float values (#1649)
An error occurred when doing calculations with two different types.

Example:
> action set_variable money:10
> action variable_math money,+,10.5,money
> ERROR

The function number_or_variable receives a string, then checks whether
this string is a number or a variable name. This way of doing this,
crashes when the function receives a float value.
Indeed, this float value is seen as a variable name because
it was not recognised as an integer.
Therefore, I upgraded the test so that it checks whether the value
passed is an integer, a float or a variable name.

One test has been added, a new test class in tests/tuxemon/test_tools.py
called "TestVariableMoney" with a method called "test_var".
This method should try all kinds of calculations possible.