Make early game more fun/balanced
Getting stuck in battles with endless healing I can't run away from in the first route isn't my idea of fun, nor is sending 5 balls in a row to catch a low health caterpillar that spams healing. Earliest game mons should be easy to catch and not nearly all of them insta-heal themselves. And whyyy does Cardiling have a power 3 murder move at level 2? The balancing here is really all over the place, this fixes the worst of it.