Fix game uncontinuable condition if buy more than two potions at start of Xeno
Check for has potion, not has exactly two potions, because players can buy from the shop, which would cause the old script to never trigger and then the player would be unable to progress in the game unless they guessed they needed to toss away all but two potions. Not my idea of fun!