Couple tweaks
Adding a "..." as Billie pauses, to make it clear she's thinking.
And I've tucked the barmaid away out of reach until she's able to be spoken to - otherwise I'm worried players will try her once, not be able to talk to her, and not try again. Same idea, I've stopped her from changing direction so it's clear she's a service provider.